Move networkinterceptor to correct subfolder devtools
